+++ backend/src/main/java/com/flat/backend/common/config/SecurityConfig.java
@@ -36,7 +36,7 @@ public class SecurityConfig{
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
         httpSecurity.csrf(AbstractHttpConfigurer::disable)
-                .authorizeHttpRequests(request -> request.requestMatchers("/auth/sign-up", "/auth/sign-in", "/auth/re-issue", "/auth/sign-out")
+                .authorizeHttpRequests(request -> request.requestMatchers("/auth/sign-up", "/auth/sign-in", "/auth/re-issue", "/auth/sign-out", "/user/test", "/user/upload")
                         .permitAll().anyRequest().authenticated())
                 .sessionManagement(manager -> manager.sessionCreationPolicy(STATELESS));
         httpSecurity.apply(new JwtSecurityConfig(jwtTokenProvider));
