--- backend/src/main/java/com/flat/backend/domain/User.java
@@ -1,10 +1,0 @@
-package com.flat.backend.domain;
-
-import lombok.Data;
-
-@Data
-public class User {
-
-    String firstName;
-    String lastName;
-}
