+++ backend/src/main/java/com/flat/backend/s3/service/AwsS3Service.java
@@ -1,4 +1,4 @@
-package com.flat.backend.service;
+package com.flat.backend.s3.service;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PutObjectRequest;
