+++ backend/src/main/resources/application-dev.yml
@@ -1,17 +1,38 @@
 spring:
   datasource:
-    url: jdbc:mysql://flat.ckm16xmstqog.us-east-2.rds.amazonaws.com/test
+    url: jdbc:mysql://flat-mysql.cvdq4nelxeht.ap-northeast-2.rds.amazonaws.com:3306/flat
     username: root
-    password: flat1234
-    driver-class-name: com.mysql.jdbc.Driver
+    password: ssafy1234
+    driver-class-name: com.mysql.cj.jdbc.Driver
 
   jpa:
     hibernate:
-      ddl-auto: create-drop
+      ddl-auto: none
     show_sql: true
     properties:
       hibernate:
         format_sql: true
 
+  servlet:
+    multipart:
+      enabled: true
+      file-size-threshold: 2MB
+      max-file-size: 5MB
+      max-request-size: 10MB
+
 server:
   port: 8000
+
+cloud:
+  aws:
+    credentials:
+      accessKey: AKIA4X5JVVJB3POTFPYN
+      secretKey: 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
+    region:
+      static: ap-northeast-2
+    stack:
+      auto: false
+
+application:
+  bucket:
+    name: flat-bucket
