+++ backend/src/main/resources/application-local.yml
@@ -3,7 +3,7 @@ spring:
     url: jdbc:mysql://127.0.0.1:3305/test
     username: root
     password: 1234
-    driver-class-name: com.mysql.jdbc.Driver
+    driver-class-name: com.mysql.cj.jdbc.Driver
 
   jpa:
     hibernate:
